refactor: order와 table의 양방향 의존성 제거
